requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1.post1
tensorflow-cpu==2.15.0
gunicorn==21.2.0
python-dotenv==1.0.1
requests==2.31.0
Werkzeug==3.0.1
joblib==1.3.2
scipy==1.12.0 
googletrans==4.0.0-rc1
